#91 - make emit statements accessible to Hyperledger
