+++ src/main/java/blf/blockchains/ethereum/EthereumListener.java
@@ -502,127 +502,6 @@ public class EthereumListener extends BaseBlockchainListener {
     }
 
     @Override
-    public void exitEmitStatementLog(BcqlParser.EmitStatementLogContext ctx) {
-        this.handleEthqlElement(ctx, this::handleEmitStatementLog);
-    }
-
-    private void handleEmitStatementLog(BcqlParser.EmitStatementLogContext ctx) {
-        final List<ValueAccessorSpecification> accessors = new LinkedList<>();
-        try {
-            for (BcqlParser.ValueExpressionContext valEx : ctx.valueExpression()) {
-                accessors.add(this.getValueAccessor(valEx));
-            }
-            final LogLineExportSpecification spec = LogLineExportSpecification.ofValues(accessors);
-            this.composer.addInstruction(spec);
-        } catch (BuildException e) {
-            LOGGER.severe(String.format("Handling of emit statement for Log files failed: %s", e.getMessage()));
-            System.exit(1);
-        }
-    }
-
-    @Override
-    public void exitEmitStatementCsv(BcqlParser.EmitStatementCsvContext ctx) {
-        this.handleEthqlElement(ctx, this::handleEmitStatementCsv);
-    }
-
-    private void handleEmitStatementCsv(BcqlParser.EmitStatementCsvContext ctx) {
-        LinkedList<CsvColumnSpecification> columns = new LinkedList<>();
-        try {
-            for (BcqlParser.NamedEmitVariableContext varCtx : ctx.namedEmitVariable()) {
-                final String name = varCtx.valueExpression().variableName() == null
-                    ? varCtx.variableName().getText()
-                    : varCtx.valueExpression().variableName().getText();
-                final ValueAccessorSpecification accessor = this.getValueAccessor(varCtx.valueExpression());
-                columns.add(CsvColumnSpecification.of(name, accessor));
-            }
-            this.composer.addInstruction(CsvExportSpecification.of(this.getValueAccessor(ctx.tableName), columns));
-        } catch (BuildException e) {
-            LOGGER.severe(String.format("Handling of emit statement for CSV files failed: %s", e.getMessage()));
-            System.exit(1);
-        }
-    }
-
-    @Override
-    public void exitEmitStatementXesTrace(BcqlParser.EmitStatementXesTraceContext ctx) {
-        this.handleEthqlElement(ctx, this::handleEmitStatementXesTrace);
-    }
-
-    private void handleEmitStatementXesTrace(BcqlParser.EmitStatementXesTraceContext ctx) {
-        try {
-            final ValueAccessorSpecification pid = this.getXesId(ctx.pid);
-            final ValueAccessorSpecification piid = this.getXesId(ctx.piid);
-            final List<XesParameterSpecification> parameters = this.getXesParameters(ctx.xesEmitVariable());
-            this.composer.addInstruction(XesExportSpecification.ofTraceExport(pid, piid, parameters));
-        } catch (BuildException e) {
-            LOGGER.severe(String.format("Handling of emit statement for XES trace files failed: %s", e.getMessage()));
-            System.exit(1);
-        }
-    }
-
-    @Override
-    public void exitEmitStatementXesEvent(BcqlParser.EmitStatementXesEventContext ctx) {
-        this.handleEthqlElement(ctx, this::handleEmitStatementXesEvent);
-    }
-
-    private void handleEmitStatementXesEvent(BcqlParser.EmitStatementXesEventContext ctx) {
-        try {
-            final ValueAccessorSpecification pid = this.getXesId(ctx.pid);
-            final ValueAccessorSpecification piid = this.getXesId(ctx.piid);
-            final ValueAccessorSpecification eid = this.getXesId(ctx.eid);
-            final List<XesParameterSpecification> parameters = this.getXesParameters(ctx.xesEmitVariable());
-            this.composer.addInstruction(XesExportSpecification.ofEventExport(pid, piid, eid, parameters));
-        } catch (BuildException e) {
-            LOGGER.severe(String.format("Handling of emit statement for XES event files failed: %s", e.getMessage()));
-            System.exit(1);
-        }
-    }
-
-    private ValueAccessorSpecification getXesId(BcqlParser.ValueExpressionContext ctx) {
-        return ctx == null ? null : this.getValueAccessor(ctx);
-    }
-
-    private List<XesParameterSpecification> getXesParameters(List<BcqlParser.XesEmitVariableContext> variables) {
-        final LinkedList<XesParameterSpecification> parameters = new LinkedList<>();
-        try {
-            for (BcqlParser.XesEmitVariableContext varCtx : variables) {
-                final String name = varCtx.variableName() == null
-                    ? varCtx.valueExpression().variableName().getText()
-                    : varCtx.variableName().getText();
-                final ValueAccessorSpecification accessor = this.getValueAccessor(varCtx.valueExpression());
-                LOGGER.info(varCtx.getText());
-
-                XesParameterSpecification parameter;
-                switch (varCtx.xesTypes().getText()) {
-                    case "xs:string":
-                        parameter = XesParameterSpecification.ofStringParameter(name, accessor);
-                        break;
-                    case "xs:date":
-                        parameter = XesParameterSpecification.ofDateParameter(name, accessor);
-                        break;
-                    case "xs:int":
-                        parameter = XesParameterSpecification.ofIntegerParameter(name, accessor);
-                        break;
-                    case "xs:float":
-                        parameter = XesParameterSpecification.ofFloatParameter(name, accessor);
-                        break;
-                    case "xs:boolean":
-                        parameter = XesParameterSpecification.ofBooleanParameter(name, accessor);
-                        break;
-                    default:
-                        throw new BuildException(String.format("Xes type '%s' not supported", varCtx.xesTypes().getText()));
-                }
-                parameters.add(parameter);
-            }
-
-            return parameters;
-        } catch (BuildException e) {
-            LOGGER.severe(String.format("Getter of XES parameters failed: %s", e.getMessage()));
-            System.exit(1);
-            return null;
-        }
-    }
-
-    @Override
     public void exitMethodStatement(BcqlParser.MethodStatementContext ctx) {
         this.handleEthqlElement(ctx, this::handleMethodStatement);
     }
@@ -694,8 +573,6 @@ public class EthereumListener extends BaseBlockchainListener {
         }
     }
 
-    // #region Utils
-
     private <T extends ParserRuleContext> void handleEthqlElement(T ctx, BuilderMethod<T> builderMethod) {
         if (this.containsError()) {
             return;
@@ -712,101 +589,5 @@ public class EthereumListener extends BaseBlockchainListener {
     private interface BuilderMethod<T> {
         void build(T ctx) throws BuildException;
     }
-
-    private ValueAccessorSpecification getValueAccessor(BcqlParser.ValueExpressionContext ctx) {
-        try {
-            if (ctx.variableName() != null) {
-                return ValueAccessorSpecification.ofVariable(ctx.getText());
-            } else if (ctx.literal() != null) {
-                return this.getLiteral(ctx.literal());
-            } else {
-                throw new UnsupportedOperationException("This value accessor specification is not supported.");
-            }
-        } catch (UnsupportedOperationException e) {
-            LOGGER.severe(String.format("Getter of value accessor failed: %s", e.getMessage()));
-            System.exit(1);
-            return null;
-        }
-    }
-
-    private ValueAccessorSpecification getLiteral(BcqlParser.LiteralContext ctx) {
-        String type = this.determineLiteralType(ctx);
-        return this.getLiteral(type, ctx.getText());
-    }
-
-    private String determineLiteralType(BcqlParser.LiteralContext ctx) {
-        try {
-            String type = null;
-            if (ctx.BOOLEAN_LITERAL() != null) {
-                type = TypeUtils.BOOL_TYPE_KEYWORD;
-            } else if (ctx.BYTES_LITERAL() != null) {
-                type = TypeUtils.BYTES_TYPE_KEYWORD;
-            } else if (ctx.INT_LITERAL() != null) {
-                type = TypeUtils.INT_TYPE_KEYWORD;
-            } else if (ctx.STRING_LITERAL() != null) {
-                type = TypeUtils.STRING_TYPE_KEYWORD;
-            } else if (ctx.arrayLiteral() != null) {
-                if (ctx.arrayLiteral().booleanArrayLiteral() != null) {
-                    type = TypeUtils.toArrayType(TypeUtils.BOOL_TYPE_KEYWORD);
-                } else if (ctx.arrayLiteral().bytesArrayLiteral() != null) {
-                    type = TypeUtils.toArrayType(TypeUtils.BYTES_TYPE_KEYWORD);
-                } else if (ctx.arrayLiteral().intArrayLiteral() != null) {
-                    type = TypeUtils.toArrayType(TypeUtils.INT_TYPE_KEYWORD);
-                } else if (ctx.arrayLiteral().stringArrayLiteral() != null) {
-                    type = TypeUtils.toArrayType(TypeUtils.STRING_TYPE_KEYWORD);
-                }
-            }
-
-            if (type == null) {
-                throw new BuildException(String.format("Cannot determine type for literal %s.", ctx.getText()));
-            }
-            return type;
-        } catch (BuildException e) {
-            LOGGER.severe(String.format("Determination of literal type failed: %s", e.getMessage()));
-            System.exit(1);
-            return null;
-        }
-
-    }
-
-    private ValueAccessorSpecification getLiteral(String type, String literal) {
-        try {
-            if (TypeUtils.isArrayType(type)) {
-                if (TypeUtils.isArrayType(type, TypeUtils.ADDRESS_TYPE_KEYWORD)) {
-                    return ValueAccessorSpecification.addressArrayLiteral(literal);
-                } else if (TypeUtils.isArrayType(type, TypeUtils.BOOL_TYPE_KEYWORD)) {
-                    return ValueAccessorSpecification.booleanArrayLiteral(literal);
-                } else if (TypeUtils.isArrayType(type, TypeUtils.BYTES_TYPE_KEYWORD)) {
-                    return ValueAccessorSpecification.bytesArrayLiteral(literal);
-                } else if (TypeUtils.isArrayType(type, TypeUtils.INT_TYPE_KEYWORD)) {
-                    return ValueAccessorSpecification.integerArrayLiteral(literal);
-                } else if (TypeUtils.isArrayType(type, TypeUtils.STRING_TYPE_KEYWORD)) {
-                    return ValueAccessorSpecification.stringArrayLiteral(literal);
-                } else {
-                    throw new BuildException(String.format("Unsupported type: '%s'.", type));
-                }
-            } else {
-                if (TypeUtils.isAddressType(type)) {
-                    return ValueAccessorSpecification.addressLiteral(literal);
-                } else if (TypeUtils.isBooleanType(type)) {
-                    return ValueAccessorSpecification.booleanLiteral(literal);
-                } else if (TypeUtils.isBytesType(type)) {
-                    return ValueAccessorSpecification.bytesLiteral(literal);
-                } else if (TypeUtils.isIntegerType(type)) {
-                    return ValueAccessorSpecification.integerLiteral(literal);
-                } else if (TypeUtils.isStringType(type)) {
-                    return ValueAccessorSpecification.stringLiteral(literal);
-                } else {
-                    throw new BuildException(String.format("Unsupported type: '%s'.", type));
-                }
-            }
-        } catch (BuildException e) {
-            LOGGER.severe(String.format("Getter of literal failed: %s", e.getMessage()));
-            System.exit(1);
-            return null;
-        }
-    }
-
-    // #endregion Utils
 
 }
